+++ spark/models.py
@@ -3,10 +3,15 @@ from django.utils.translation import ugettext_lazy as _
 
 
 class EventQuerySet(models.QuerySet):
-    def create_if_new(self, **kwargs):
+    def create_if_new(self, event):
+        context = dict(event)
         try:
             with transaction.atomic():
-                return self.create(**kwargs)
+                return self.create(
+                    key=context.pop("key"),
+                    group=context.pop("group"),
+                    context=repr(context),
+                )
         except IntegrityError:
             return None